landing: rename countryMessageBoardView -> messageBoardPanel
