+++ webapp/app/appView.js
@@ -12,7 +12,7 @@ import CountrySelection from '@webapp/app/components/countrySelection'
 import Navigation from '@webapp/app/components/navigation/navigation'
 import Review from '@webapp/app/assessment/components/review/review'
 import UserChat from '@webapp/app/user/chat/userChatView'
-import CountryMessageBoardView from '@webapp/app/landing/views/messageBoard/countryMessageBoardView'
+import MessageBoardPanel from '@webapp/app/landing/views/messageBoard/messageBoardPanel'
 import ErrorComponent from '@webapp/app/components/error/errorComponent'
 import AssessmentPrintView from '@webapp/app/assessment/components/print/assessmentPrintView'
 
@@ -63,7 +63,7 @@ const LoggedInView = () => {
           <>
             <Review />
             <UserChat />
-            <CountryMessageBoardView />
+            <MessageBoardPanel />
           </>
         )}
         <div className="app-view">
